Extract game engine, player identity, and HUD from Game.tsx
Game.tsx is now a thin shell exposing the Jazz contract types
(GameState, RemotePlayer) and bridging engine state to onStateChange.
Non-Jazz concerns moved to game/engine.ts, game/player.ts, game/Hud.tsx.
